feat(create_pair): Instantiate2, doctest for example
Added Instantiate2 for creation of an LP token. If we want to use Token factory we can refactor but this covers the general OR case. Thanks to @kerber0x for help with instantiate2

Cleaned up around the function and added some more messages, errors and such.
There is a doctest for people who want to just call this func. This should be enhanced later to show with messages