fix(testing): reorganize testing utilities
Moves the Osmosis mocking to the `white-whale-testing` package, rather
than `white-whale-std` package. This lets us drop the `#[cfg(test)]`
qualifiers which restrict integration tests from consuming our package.

This is a suitable restriction to drop, since it is likely that any code
that is using `white-whale-testing` is performing testing operations,
while our `white-whale-std` was public facing.